Refactor: payment & ShoppingCart & Like Controller 로직 생성[박한솔]
