+++ basilium-server/src/main/java/basilium/basiliumserver/controller/payment/PaymentController.java
@@ -1,8 +1,12 @@
 package basilium.basiliumserver.controller.payment;
 
-import basilium.basiliumserver.domain.purchaseTransaction.kafkaPaymentInventory.PaymentInventoryResponse;
-import basilium.basiliumserver.domain.purchaseTransaction.kafkaPaymentInventory.ProductUpdateMessage;
-import basilium.basiliumserver.service.purchaseTransaction.PurchaseTransactionService;
+import basilium.basiliumserver.auth.support.AuthUser;
+import basilium.basiliumserver.domain.payment.OrderPaymentRequest;
+import basilium.basiliumserver.domain.payment.kafkaPaymentInventory.PaymentInventoryResponse;
+import basilium.basiliumserver.domain.payment.kafkaPaymentInventory.ProductUpdateMessage;
+import basilium.basiliumserver.domain.user.NormalUser;
+import basilium.basiliumserver.service.payment.PaymentService;
+import basilium.basiliumserver.service.user.NormalUserService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -10,6 +14,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -21,11 +26,15 @@ import java.util.concurrent.ConcurrentHashMap;
 @Slf4j
 public class PaymentController {
 
+    private final PaymentService paymentService;
+    private final NormalUserService normalUserService;
+
+
     private final KafkaTemplate<String, String> kafkaTemplate;
-    private final PurchaseTransactionService purchaseTransactionService;
     private final ObjectMapper objectMapper = new ObjectMapper();
     private static final ConcurrentHashMap<UUID, String> requestTaskMaps = new ConcurrentHashMap<>();
-    //private final ConcurrentHashMap<String, String> requestTaskMap = new ConcurrentHashMap<>();
+
+
     @PostMapping("/request")
     public ResponseEntity<PaymentInventoryResponse> requestPayment(@RequestParam Long productId, @RequestParam Long count) throws Exception {
         UUID requestId = UUID.randomUUID();
@@ -43,7 +52,7 @@ public class PaymentController {
 
         // Call the service method and capture the PaymentInventoryResponse object
         //서비스 사용안하게 리펙토링하기
-        PaymentInventoryResponse response = purchaseTransactionService.scheduleRestoration(productId, count, requestId);
+        PaymentInventoryResponse response = paymentService.scheduleRestoration(productId, count, requestId);
         return ResponseEntity.ok(response);
 
         //밑에는 서비스를 사용안해도 되는거 테스트
@@ -57,51 +66,7 @@ public class PaymentController {
 
          */
     }
-    /*
-    //타이머랑 함께 리턴
-    @PostMapping("/request")
-    public ResponseEntity<String> requestPayment(@RequestParam Long productId, @RequestParam Long count) throws Exception {
-        UUID requestId = UUID.randomUUID(); // 요청 ID를 UUID로 생성
 
-        ProductUpdateMessage message = new ProductUpdateMessage(productId, count, requestId);
-
-        String messageStr = objectMapper.writeValueAsString(message);
-        kafkaTemplate.send("product-update-topic", messageStr);
-
-        // 스케줄된 작업을 추가하고 그 정보를 반환받음
-        String responseMessage = purchaseTransactionService.scheduleRestoration(productId, count, requestId);
-
-        String key = productId + "-" + count; // productId와 count를 조합하여 고유한 키 생성
-        requestTaskMaps.put(requestId, key); // (productId+count)key와 requestId를 사용하여 매핑하여 저장
-        log.info("************************************************");
-        log.info("requestId {} with key {}", requestId, key);
-
-        // 클라이언트에게 요청 ID와 스케줄링된 지연 시간을 포함한 응답을 반환
-        return ResponseEntity.ok(responseMessage);
-    }
-
-     */
-    /*
-    //원본
-    @PostMapping("/request")
-    public ResponseEntity<String> requestPayment(@RequestParam Long productId, @RequestParam Long count) throws Exception {
-        UUID requestId = UUID.randomUUID(); // 요청 ID를 UUID로 생성
-
-        ProductUpdateMessage message = new ProductUpdateMessage(productId, count, requestId);
-
-        String messageStr = objectMapper.writeValueAsString(message);
-        kafkaTemplate.send("product-update-topic", messageStr);
-
-        purchaseTransactionService.scheduleRestoration(productId, count, requestId);
-
-        String key = productId + "-" + count; // productId와 count를 조합하여 고유한 키 생성
-        requestTaskMaps.put(requestId, key); // (productId+count)key와 requestId를 사용하여 매핑하여 저장
-        log.info("************************************************");
-        log.info("requestId {} with key {}",requestId, key);
-        return ResponseEntity.ok(requestId.toString());
-    }
-
-     */
 
 
     @PostMapping("/response")
@@ -123,7 +88,7 @@ public class PaymentController {
 
         UUID taskId = requestId;
         // requestId를 그대로 taskId로 사용하여 processPaymentResponse 호출
-        purchaseTransactionService.processPaymentResponse(productId, count, success, taskId);
+        paymentService.processPaymentResponse(productId, count, success, taskId);
 
         // 요청이 처리된 후에는 해당 매핑을 제거하여 메모리 누수를 방지합니다.
         log.info("[컨트롤러-response]:requestTaskMaps 메모리 해제 전: Getting request task maps - size: {}", requestTaskMaps.size());
@@ -138,5 +103,22 @@ public class PaymentController {
         log.info("[컨트롤러-request]:requestTaskMaps 메모리 해제 후: Getting request task maps - size: {}", requestTaskMaps.size());
     }
 
+    @GetMapping("/order/history")
+    public ResponseEntity<List<?>> userOrderInfos(@AuthUser String userId){
+        NormalUser ret = normalUserService.userInfoById(userId);
+        return ResponseEntity.ok(paymentService.userOrderHistory(ret.getUserNumber()));
+    }
+
+    @GetMapping("/order/payment")
+    public ResponseEntity<?> saveUserPayment(@AuthUser String userId){
+        NormalUser ret = normalUserService.userInfoById(userId);
+        return ResponseEntity.ok(paymentService.userOrderHistory(ret.getUserNumber()));
+    }
+
+    @PostMapping("order/payment/{impUid}")
+    public ResponseEntity<String> handlePayment(@AuthUser String userId, @PathVariable String impUid, @RequestBody OrderPaymentRequest request) {
+        paymentService.processPayment(userId, impUid, request);
+        return ResponseEntity.ok("Payment processed successfully");
+    }
 }
