--- basilium-server/src/main/java/basilium/basiliumserver/order/OrderStatus.java
@@ -1,6 +1,0 @@
-package basilium.basiliumserver.order;
-
-public enum OrderStatus {
-    PAID_UP,
-    PAYMENT_FAILED
-}
