+++ basilium-server/src/main/java/basilium/basiliumserver/repository/shoppingCart/JpaShoppingCartRepo.java
@@ -1,7 +1,7 @@
 package basilium.basiliumserver.repository.shoppingCart;
 
 
-import basilium.basiliumserver.domain.purchaseTransaction.OrderListDAO;
+import basilium.basiliumserver.domain.payment.OrderListDAO;
 import basilium.basiliumserver.domain.shoppingCart.ShoppingCart;
 import jakarta.persistence.EntityManager;
 import java.util.List;
